# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon Backend API
  version: 1.1.0
paths:
...
  /admin/reports/overview:
    get:
      summary: Get admin analytics overview for a range
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          required: true
          schema:
            type: string
            enum: [24h, 7d, 30d]
        - name: timezone
          in: query
          required: false
          schema:
            type: string
            default: Asia/Manila
      responses:
        "200":
          description: Reports overview payload
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Missing or invalid admin bearer token
        "500":
          $ref: "#/components/responses/Error"
  /admin/reports/generate:
    post:
      summary: Generate and persist admin report run
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [report_key]
              properties:
                report_key:
                  type: string
                  enum:
                    - daily_safety_report
                    - weekly_safety_report
                    - monthly_safety_report
                timezone:
                  type: string
                  default: Asia/Manila
      responses:
        "200":
          description: Reports overview payload after generation
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Missing or invalid admin bearer token
        "500":
          $ref: "#/components/responses/Error"
  /admin/reports/export.csv:
    get:
      summary: Export admin analytics as CSV
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          required: true
          schema:
            type: string
            enum: [24h, 7d, 30d]
        - name: timezone
          in: query
          required: false
          schema:
            type: string
            default: Asia/Manila
      responses:
        "200":
          description: CSV download
          content:
            text/csv:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Missing or invalid admin bearer token
        "500":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Error:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"